+++ ShellSort/shellsort_kn.cpp
@@ -50,7 +50,7 @@ void GenerateGaps(const size_t &n, std::vector<long long> &gaps)
 	return;
 }
 
-void RunShellSort(std::vector<float> &arr, const size_t &n)
+void RunShellSort(std::vector<double> &arr, const size_t &n)
 {
 	std::vector<long long> gaps;
 	GenerateGaps(n, gaps);
@@ -60,7 +60,7 @@ void RunShellSort(std::vector<float> &arr, const size_t &n)
 	for (size_t i = 0; i < gaps_count; ++i) {
 		long long gap = gaps[i];
 		for (size_t j = gap; j < n; ++j) {
-			float temp = arr[j];
+			double temp = arr[j];
 			long long k;
 			for (k = j; k >= gap && arr[k - gap] > temp; k -= gap) {
 				arr[k] = arr[k - gap];
@@ -72,7 +72,7 @@ void RunShellSort(std::vector<float> &arr, const size_t &n)
 	return;
 }
 
-void ReadArray(std::vector<float> &arr, size_t &n, char *input_file)
+void ReadArray(std::vector<double> &arr, size_t &n, char *input_file)
 {
 	std::ifstream in(input_file);
 
@@ -88,7 +88,7 @@ void ReadArray(std::vector<float> &arr, size_t &n, char *input_file)
 	return;
 }
 
-void PrintArray(std::vector<float> &arr, size_t &n, char *output_file)
+void PrintArray(std::vector<double> &arr, size_t &n, char *output_file)
 {
 	std::ofstream out(output_file);
 
@@ -109,7 +109,7 @@ int main(int argc, char *argv[])
 	}
 
 	size_t n;
-	std::vector<float> array;
+	std::vector<double> array;
 
 	ReadArray(array, n, argv[1]);
